+++ app/src/main/java/com/gitspark/gitspark/api/model/ApiRepo.kt
@@ -12,7 +12,7 @@ data class ApiStarredRepo(
 }
 
 data class ApiRepo(
-    @field:Json(name = "id") val id: Int?,
+    @field:Json(name = "id") val id: Long?,
     @field:Json(name = "name") val name: String?,
     @field:Json(name = "full_name") val fullName: String?,
     @field:Json(name = "owner") val owner: ApiUser?,
