+++ app/src/main/java/com/gitspark/gitspark/model/Token.kt
@@ -3,7 +3,7 @@ package com.gitspark.gitspark.model
 const val PREFERENCES_TOKEN = "PREFERENCES_TOKEN"
 
 data class Token(
-    val tokenId: Int = 0,
+    val tokenId: Long = 0,
     val value: String = "",
     val hashedValue: String = "",
     val scopes: List<String> = emptyList(),
